backend/requirements.txt
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.99.1
fastapi-jwt-auth==0.5.0
h11==0.14.0
idna==3.6
MarkupSafe==2.1.3
minio==7.2.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.11
PyJWT==1.7.1
PyMySQL==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
Werkzeug==3.0.1
